[Refactor]: Change data type of multiple vector to size_t from int

+++ Nexus/src/ECS/Component.cpp
@@ -2,4 +2,4 @@
 #include "Component.h"
 
 /// Allocating memory for the static variable
-int IComponent::m_nextId = 0;
+size_t IComponent::m_nextId = 0;
